# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * JEE Planner API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TaskPriority = typeof TaskPriority[keyof typeof TaskPriority];


export const TaskPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface Task {
  id: number;
  title: string;
  subject: string;
  /** @nullable */
  chapter?: string | null;
  /** Date in YYYY-MM-DD format */
  dueDate: string;
  priority: TaskPriority;
  completed: boolean;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export type TaskInputPriority = typeof TaskInputPriority[keyof typeof TaskInputPriority];


export const TaskInputPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  subject: string;
  chapter?: string;
  dueDate: string;
  priority: TaskInputPriority;
  notes?: string;
}

export type TaskUpdatePriority = typeof TaskUpdatePriority[keyof typeof TaskUpdatePriority];


export const TaskUpdatePriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface TaskUpdate {
  /** @minLength 1 */
  title?: string;
  subject?: string;
  /** @nullable */
  chapter?: string | null;
  dueDate?: string;
  priority?: TaskUpdatePriority;
  completed?: boolean;
  /** @nullable */
  notes?: string | null;
}

export type TestType = typeof TestType[keyof typeof TestType];


export const TestType = {
  coaching: 'coaching',
  school: 'school',
  mock: 'mock',
  jee: 'jee',
} as const;

export interface Test {
  id: number;
  name: string;
  /** Date in YYYY-MM-DD format */
  date: string;
  /** Time in HH:MM format */
  time: string;
  type: TestType;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export type TestInputType = typeof TestInputType[keyof typeof TestInputType];


export const TestInputType = {
  coaching: 'coaching',
  school: 'school',
  mock: 'mock',
  jee: 'jee',
} as const;

export interface TestInput {
  /** @minLength 1 */
  name: string;
  date: string;
  time: string;
  type: TestInputType;
  notes?: string;
}

export type TestUpdateType = typeof TestUpdateType[keyof typeof TestUpdateType];


export const TestUpdateType = {
  coaching: 'coaching',
  school: 'school',
  mock: 'mock',
  jee: 'jee',
} as const;

export interface TestUpdate {
  /** @minLength 1 */
  name?: string;
  date?: string;
  time?: string;
  type?: TestUpdateType;
  /** @nullable */
  notes?: string | null;
}

export type SettingsTheme = typeof SettingsTheme[keyof typeof SettingsTheme];


export const SettingsTheme = {
  dark: 'dark',
  light: 'light',
} as const;

export interface Settings {
  id: number;
  theme: SettingsTheme;
  accentColor: string;
  jeeMainDate: string;
  jeeAdvancedDate: string;
}

export type SettingsUpdateTheme = typeof SettingsUpdateTheme[keyof typeof SettingsUpdateTheme];


export const SettingsUpdateTheme = {
  dark: 'dark',
  light: 'light',
} as const;

export interface SettingsUpdate {
  theme?: SettingsUpdateTheme;
  accentColor?: string;
  jeeMainDate?: string;
  jeeAdvancedDate?: string;
}

export interface DashboardSummary {
  todayTasks: Task[];
  upcomingTests: Test[];
  weekTasks: Task[];
  /** @nullable */
  daysToJeeMain: number | null;
  /** @nullable */
  daysToJeeAdvanced: number | null;
  todayDate: string;
}

export type ListTasksParams = {
/**
 * Filter by due date (YYYY-MM-DD)
 */
date?: string;
/**
 * Filter by subject
 */
subject?: string;
/**
 * Filter by completion status
 */
completed?: boolean;
/**
 * Search query
 */
search?: string;
};

export type ListTestsParams = {
/**
 * Search query
 */
search?: string;
/**
 * Only upcoming tests
 */
upcoming?: boolean;
};