# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * JEE Planner API
 * OpenAPI spec version: 0.1.0
 */
import type { Task } from './task';
import type { Test } from './test';

export interface DashboardSummary {
  todayTasks: Task[];
  upcomingTests: Test[];
  weekTasks: Task[];
  /** @nullable */
  daysToJeeMain: number | null;
  /** @nullable */
  daysToJeeAdvanced: number | null;
  todayDate: string;
}