# lib/api-zod/src/generated/types/settings.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * JEE Planner API
 * OpenAPI spec version: 0.1.0
 */
import type { SettingsTheme } from './settingsTheme';

export interface Settings {
  id: number;
  theme: SettingsTheme;
  accentColor: string;
  jeeMainDate: string;
  jeeAdvancedDate: string;
}