# lib/api-zod/src/generated/types/settingsUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * JEE Planner API
 * OpenAPI spec version: 0.1.0
 */
import type { SettingsUpdateTheme } from './settingsUpdateTheme';

export interface SettingsUpdate {
  theme?: SettingsUpdateTheme;
  accentColor?: string;
  jeeMainDate?: string;
  jeeAdvancedDate?: string;
}