# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * JEE Planner API
 * OpenAPI spec version: 0.1.0
 */
import type { TaskPriority } from './taskPriority';

export interface Task {
  id: number;
  title: string;
  subject: string;
  /** @nullable */
  chapter?: string | null;
  /** Date in YYYY-MM-DD format */
  dueDate: string;
  priority: TaskPriority;
  completed: boolean;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}