# lib/api-zod/src/generated/types/taskInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * JEE Planner API
 * OpenAPI spec version: 0.1.0
 */
import type { TaskInputPriority } from './taskInputPriority';

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  subject: string;
  chapter?: string;
  dueDate: string;
  priority: TaskInputPriority;
  notes?: string;
}