# lib/api-zod/src/generated/types/testInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * JEE Planner API
 * OpenAPI spec version: 0.1.0
 */
import type { TestInputType } from './testInputType';

export interface TestInput {
  /** @minLength 1 */
  name: string;
  date: string;
  time: string;
  type: TestInputType;
  notes?: string;
}