# lib/api-zod/src/generated/types/testUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * JEE Planner API
 * OpenAPI spec version: 0.1.0
 */
import type { TestUpdateType } from './testUpdateType';

export interface TestUpdate {
  /** @minLength 1 */
  name?: string;
  date?: string;
  time?: string;
  type?: TestUpdateType;
  /** @nullable */
  notes?: string | null;
}